Extract API course data to module
